Test Wobble Animation`
A small experiment once again in the Game Scene Mock-Up where I added an animation to the two node slimes, making them wobble. This was a simple expirement in what the animator in Unity feels like. Mostly likely will not be used for animations, but might be used for placeholders.