Remake Scene for Testing
This was the first incident of facing true issues with the unity meta files. I'm not positive how I broke them, but they became impossible to fix. I was forced to delete GameSceneMockUp and had to just remake the elements in sample scene. Will have to be more careful from here on out.